Finished chapter 5
Just spent around an hour tracking a bug that was causing the sphere
to render too large. I checked every single method on Vec3, and not dice.

After some more investigation on the dot product method, I finally noticed
that the last set of multiplications was actually an addition on my code.

Name a better way to spen an hour...